+++ hw3_EM.py
@@ -138,7 +138,7 @@ infer_act_prob = {'North': np.array([[0.0, 1.0, 0.0, 0.0, 0.0],
                                     )
                  }
 
-grid_map = np.array(range(0,6)).reshape([2,3])
+grid_map = np.array(range(0,6), dtype=np.int8).reshape([2,3])
 
 shortest_paths = {frozenset([0, 1]): (0, 1),
                   frozenset([0, 2]): (0, 1, 2),
